refactor: divide mapper
